requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
requests==2.31.0
email-validator==2.0.0
reportlab
gunicorn
flask_migrate
psycopg2-binary
pymongo